Move statsd cleaner into library
